requirements.txt
# LOVE Protocol Phase 0 Lab Environment
# AI Simulation Framework

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Ollama client
ollama>=0.1.0

# Embeddings and NLP
sentence-transformers>=2.2.0
torch>=2.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# Optional: for advanced embeddings
# transformers>=4.30.0

